Struct particle::automatons::NFA

pub struct NFA {
    pub initial_state: StateId,
    pub final_states: FxHashMap<StateId, BranchId>,
    pub transitions: MultiMap<(StateId, Transition), StateId>,
}

Non-deterministic Finite Automaton.

The inside implementation of the automaton is based on u8, therefore a character transition may be represented as multiple edges in the NFA depending on its UTF-8 encoding.

Fields
initial_state: StateId final_states: FxHashMap<StateId, BranchId> transitions: MultiMap<(StateId, Transition), StateId>
Methods

impl NFA

pub fn new() -> Self

Constructs an empty NFA.

pub fn transition_set(&self, from: &StateSet, input: u8) -> StateSet

Calculates the transition set of a stateset with given input.

pub fn set_branch(&mut self, branch: BranchId)

Sets the branch id for all final states currently in the NFA.

This should only be called right before you convert the NFA into DFA, without any further operation after the call except conversion, or the branch information will be a total mess!

pub fn zero_or_more(self) -> NFA

Repeats self by >=0 times (* in regex).

pub fn one_or_more(self) -> NFA

Repeats self by >=1 times (+ in regex).

pub fn optional(&self) -> NFA

Makes self optional (0/1 times).